requirements.txt
blinker==1.9.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.3
pandas==2.3.3
pillow==11.3.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.4
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
xlsxwriter==3.2.9
cryptography==44.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
pywebview==5.3
